refactor(jans-auth-server): split registration into multiple actions/services
https://github.com/JanssenProject/jans/issues/1126
